+++ src/app/src/components/messageTypes/card/GroupInvite.tsx
@@ -6,7 +6,7 @@ import {
 } from '@shared/redux/actions/group';
 import { getGroupInviteInfoCache } from '@src/shared/utils/cache-helper';
 import _get from 'lodash/get';
-import { MessageProps } from '@src/shared/components/MessageHandler';
+import { MessageProps } from '@shared/components/message/MessageHandler';
 import { TRPGState } from '@redux/types/__all__';
 
 // 入团邀请
