+++ src/web/components/messageTypes/card/DiceInvite.tsx
@@ -1,7 +1,7 @@
 import BaseCard from './BaseCard';
 import { connect, DispatchProp } from 'react-redux';
 import { acceptDiceInvite } from '../../../../shared/redux/actions/dice';
-import { MessageProps } from '@shared/components/MessageHandler';
+import { MessageProps } from '@shared/components/message/MessageHandler';
 import { TRPGState } from '@src/shared/redux/types/__all__';
 
 interface Props extends MessageProps, DispatchProp<any> {
