+++ src/web/components/messageTypes/card/GroupInvite.tsx
@@ -4,7 +4,7 @@ import {
   agreeGroupInvite,
   refuseGroupInvite,
 } from '@shared/redux/actions/group';
-import { MessageProps } from '@shared/components/MessageHandler';
+import { MessageProps } from '@shared/components/message/MessageHandler';
 import { TRPGState } from '@src/shared/redux/types/__all__';
 import { getGroupInviteInfoCache } from '@shared/utils/cache-helper';
 import _get from 'lodash/get';
