+++ src/web/components/messageTypes/card/GroupRequest.tsx
@@ -5,7 +5,7 @@ import {
   refuseGroupRequest,
 } from '@shared/redux/actions/group';
 import { TRPGState } from '@redux/types/__all__';
-import { MessageProps } from '@shared/components/MessageHandler';
+import { MessageProps } from '@shared/components/message/MessageHandler';
 import _isNil from 'lodash/isNil';
 
 interface Props extends MessageProps, DispatchProp<any> {
